+++ packages/expo-audio/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### 🐛 Bug fixes
 
 - [iOS] Support base64 strings as an audio source. ([#37031](https://github.com/expo/expo/pull/37031) by [@alanjhughes](https://github.com/alanjhughes))
+- [iOS] Correctly add the http headers to the `AVURLAsset`.
 
 ### 💡 Others